Course Test Part 1

+++ modules/Cyaxaress/Category/Http/Controllers/CategoryController.php
@@ -23,12 +23,14 @@ class CategoryController extends Controller
 
     public function store(CategoryRequest $request)
     {
+        $this->authorize('manage', Category::class);
         $this->repo->store($request);
         return back();
     }
 
     public function edit($categoryId)
     {
+        $this->authorize('manage', Category::class);
         $category = $this->repo->findById($categoryId);
         $categories = $this->repo->allExceptById($categoryId);
         return view('Categories::edit', compact('category', 'categories'));
@@ -36,12 +38,14 @@ class CategoryController extends Controller
 
     public function update($categoryId, CategoryRequest $request)
     {
+        $this->authorize('manage', Category::class);
         $this->repo->update($categoryId, $request);
         return back();
     }
 
     public function destroy($categoryId)
     {
+        $this->authorize('manage', Category::class);
         $this->repo->delete($categoryId);
         return AjaxResponses::SuccessResponse();
     }
